Move workflows out of LambdaCreator, reduce code duplication
