+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/upload/steps/LambdaWorkflows.kt
@@ -1,0 +1,78 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package software.aws.toolkits.jetbrains.services.lambda.upload.steps
+
+import com.intellij.openapi.project.Project
+import software.aws.toolkits.jetbrains.core.awsClient
+import software.aws.toolkits.jetbrains.core.credentials.AwsConnectionManager
+import software.aws.toolkits.jetbrains.core.credentials.toEnvironmentVariables
+import software.aws.toolkits.jetbrains.services.lambda.sam.SamOptions
+import software.aws.toolkits.jetbrains.services.lambda.sam.SamTemplateUtils
+import software.aws.toolkits.jetbrains.services.lambda.upload.FunctionUploadDetails
+import software.aws.toolkits.jetbrains.utils.execution.steps.StepWorkflow
+import java.nio.file.Files
+import java.nio.file.Path
+
+fun createLambdaWorkflow(
+    project: Project,
+    functionUploadDetails: FunctionUploadDetails,
+    buildDir: Path,
+    codeUri: Path,
+    codeStorageLocation: String,
+    samOptions: SamOptions,
+    functionDetails: FunctionUploadDetails
+): StepWorkflow {
+    val (dummyTemplate, dummyLogicalId) = createTemporaryTemplate(buildDir, functionUploadDetails, codeUri)
+    val packagedTemplate = buildDir.resolve("packaged-temp-template.yaml")
+    val envVars = createAwsEnvVars(project)
+
+    return StepWorkflow(
+        BuildLambda(dummyTemplate, buildDir, samOptions),
+        PackageLambda(dummyTemplate, packagedTemplate, dummyLogicalId, codeStorageLocation, envVars),
+        CreateLambda(project.awsClient(), functionDetails)
+    )
+}
+
+fun updateLambdaCodeWorkflow(
+    project: Project,
+    functionUploadDetails: FunctionUploadDetails,
+    buildDir: Path,
+    codeUri: Path,
+    codeStorageLocation: String,
+    samOptions: SamOptions,
+    updatedFunctionDetails: FunctionUploadDetails?
+): StepWorkflow {
+    val (dummyTemplate, dummyLogicalId) = createTemporaryTemplate(buildDir, functionUploadDetails, codeUri)
+    val packagedTemplate = buildDir.resolve("packaged-temp-template.yaml")
+    val envVars = createAwsEnvVars(project)
+
+    return StepWorkflow(
+        BuildLambda(dummyTemplate, buildDir, samOptions),
+        PackageLambda(dummyTemplate, packagedTemplate, dummyLogicalId, codeStorageLocation, envVars),
+        UpdateLambdaCode(project.awsClient(), functionUploadDetails.name, updatedFunctionDetails)
+    )
+}
+
+private fun createAwsEnvVars(project: Project): Map<String, String> {
+    val connectSettings = AwsConnectionManager.getInstance(project).connectionSettings()
+        ?: throw IllegalStateException("Tried to update a lambda without valid AWS connection")
+
+    return connectSettings.credentials.resolveCredentials().toEnvironmentVariables() + connectSettings.region.toEnvironmentVariables()
+}
+
+private fun createTemporaryTemplate(buildDir: Path, functionUploadDetails: FunctionUploadDetails, codeUri: Path): Pair<Path, String> {
+    Files.createDirectories(buildDir)
+
+    val dummyTemplate = Files.createTempFile("temp-template", ".yaml")
+    val dummyLogicalId = "Function"
+    SamTemplateUtils.writeDummySamTemplate(
+        tempFile = dummyTemplate,
+        logicalId = dummyLogicalId,
+        runtime = functionUploadDetails.runtime,
+        handler = functionUploadDetails.handler,
+        codeUri = codeUri.toString()
+    )
+
+    return Pair(dummyTemplate, dummyLogicalId)
+}
